Delete save file after reading data for stats
Now deletes save file after reading the data and putting it into the saveGame object for the stats interaction. This saves a significant amount of memory, particularly in the case that somebody leaves the interaction running indefinitely. (TODO: make a time-out to delete interactions after a time so that there can't be an accumilation.)