Added LOAD command
LOAD command loads a previous reserve that was not ENDed in a previous session but is not currently open in a AbstractChannel. This basically means, if the program is closed without all Reserves being finished, those lists can be opened back up again.
This will not continue editing the same message as previously (TODO)
The discord channel needs to be the same one, not just the same name. If the channel no longer exists because it was deleted, then it's gone. (It will still be on file, but readding will be manual. TODO fix that).