Improved Tag Localizations
Wrote a script to go through localization files and find tags. Will add alternative names for "The" or accents. May include other localizations that look like tags but are not really. Also puts tags in alphabetical order.

Updated localization using the new program (and a bit of manual filtering of non-nation tag lookalikes) to fix issues with some tags not being included.