Colonial Flag fix 3 - works
Removed some of the "nat.tag"s but for some reason just removing the error handling for the main flag image display fixes it. I'm not sure if this will cause problems in the case there is an actual error, but for now it somehow seems to cause problems when there isn't one.