Optimized stats border tracing
Border tracing is the most time and resource-intensive part of the stats image creation, and seriously needed optimization. This update is a substantial improvement.
On my PC, the test file I have been using went from being about 35 seconds to only about 8 seconds for border tracing.
This was mostly done with two methods: not repeating computations which, though simple, are slow when done thousands of times and reducing calls to the Pillow library as much as possible by simply loading all the data at the beginning then accessing the local copy.