Replace stats img slicing with sending bytes to C++
Replaced the method of slicing up the Python PIL image then sending a Python array of pixels to the EU4cpplib with a method of directly sending the pixel data in a stream of bytes (33 MB), a much cheaper action due to the lack of Python datastructures such as lists and tuples.

On the python side, PIL converts the image to a bytes object which then becomes a string in C++ and can be iterated upon for each R, G, or B value that are a byte each. It seems to limit memory consumption during image generation to below 150MB while doing it all at once.