Reserve map draw optimization
Using a new system where we pregenerate a list of all the locations of the 1444 capital locations, saved in tagCapitals.txt. This is generated using CapitalLocations.py from the 1444 save file and EU4Lib.province() (which references positions.txt). This slims down the image generation time to be negligible; the only significant lag time is now upload speed.