Implemented automatic uploading to Skanderbeg.pm
Implemented a system to automatically upload to Skanderbeg.pm's API and get a link to share alongside stats.
Using asyncio and aiohttp, we avoid freezing the normal stats generation while uploading to skanderbeg, while still minimizing the time the full save file is in memory.
All the saves get uploaded to my personal profile on skanderbeg so that could be an issue. Skanderbeg's API support is also sometimes a bit unstable, so if a command is given `$stats false` we skip the skanderbeg upload.