Last map generation optimization for awhile
Map generation is now down to about half a second on my machine, though it will likely take longer on the server.

generateLandAdjacency: Getting the province ID from the province color is now done as late as possible. To accommodate this, we create a vector of the water province colors at the beginning and simply check for water provinces by color instead of by ID.